Refactored project into modules
