+++ numba_rvsdg/utils/map_utils.py
@@ -1,0 +1,4 @@
+import dis
+
+def bcmap_from_bytecode(bc: dis.Bytecode):
+    return {inst.offset: inst for inst in bc}
